let "anybody" start an X server, instead of just console users
On recent X servers, /dev/tty is not longer a "console" device when evaluating
the "only allow console users" constraint, only /dev/ttyn devices. Since the X
server is started from init, which presumably runs on /dev/tty, X refuses to
start.

Allowing anybody to start the X server fixes this. Given that there cannot be
remote users on the system, this should not impact the security much.